Remove formatting errors
