+++ packages/components/lib/index.d.ts
@@ -20,7 +20,7 @@ import PlusChip from "./PlusChip/PlusChip";
 import Progress from "./Progress/Progress";
 import Select from "./Select/Select";
 import Sidebar, { SidebarItem, SidebarLink } from "./Sidebar/Sidebar";
-import SideNavigation, { SideNavigationHeader, SideNavigationItem, SideNavigationLink } from "./SideNavigation/SideNavigation";
+import Sidenav, { SidenavHeader, SidenavItem } from "./Sidenav/Sidenav";
 import Spinner from "./Spinner/Spinner";
 import Switch from "./Switch/Switch";
 import { Tabs, Tab } from "./Tabs/Tabs";
@@ -28,4 +28,4 @@ import Tooltip from "./Tooltip/Tooltip";
 import Timeline, { TimelineItem } from "./Timeline/Timeline";
 import { TitleType, Heading1Type, Heading2Type, BodyType, SmallType } from "./Typography/Typography";
 import Upload from "./Upload/Upload";
-export { Button, ButtonGroup, Breakdown, Card, CardHeader, Chip, ColorPicker, ContextMenu, ContextMenuItem, Grid, DatePicker, Fieldset, Header, HeaderItem, HeaderSeparator, HeaderTitle, Icon, ReactFeatherIconName, InfoTile, Input, Modal, Paginator, PlusChip, Progress, Select, SideNavigation, SideNavigationHeader, SideNavigationItem, SideNavigationLink, Sidebar, SidebarItem, SidebarLink, Spinner, Switch, Tabs, Tab, Timeline, TimelineItem, Tooltip, TitleType, Heading1Type, Heading2Type, BodyType, SmallType, ThemeProvider, operationalTheme, Theme, Upload };
+export { Button, ButtonGroup, Breakdown, Card, CardHeader, Chip, ColorPicker, ContextMenu, ContextMenuItem, Grid, DatePicker, Fieldset, Header, HeaderItem, HeaderSeparator, HeaderTitle, Icon, ReactFeatherIconName, InfoTile, Input, Modal, Paginator, PlusChip, Progress, Select, Sidenav, SidenavHeader, SidenavItem, Sidebar, SidebarItem, SidebarLink, Spinner, Switch, Tabs, Tab, Timeline, TimelineItem, Tooltip, TitleType, Heading1Type, Heading2Type, BodyType, SmallType, ThemeProvider, operationalTheme, Theme, Upload };
