+++ packages/components/lib/index.js
@@ -52,11 +52,10 @@ var Sidebar_1 = require("./Sidebar/Sidebar");
 exports.Sidebar = Sidebar_1.default;
 exports.SidebarItem = Sidebar_1.SidebarItem;
 exports.SidebarLink = Sidebar_1.SidebarLink;
-var SideNavigation_1 = require("./SideNavigation/SideNavigation");
-exports.SideNavigation = SideNavigation_1.default;
-exports.SideNavigationHeader = SideNavigation_1.SideNavigationHeader;
-exports.SideNavigationItem = SideNavigation_1.SideNavigationItem;
-exports.SideNavigationLink = SideNavigation_1.SideNavigationLink;
+var Sidenav_1 = require("./Sidenav/Sidenav");
+exports.Sidenav = Sidenav_1.default;
+exports.SidenavHeader = Sidenav_1.SidenavHeader;
+exports.SidenavItem = Sidenav_1.SidenavItem;
 var Spinner_1 = require("./Spinner/Spinner");
 exports.Spinner = Spinner_1.default;
 var Switch_1 = require("./Switch/Switch");
