+++ packages/components/src/index.ts
@@ -23,10 +23,7 @@ import PlusChip from "./PlusChip/PlusChip"
 import Progress from "./Progress/Progress"
 import Select from "./Select/Select"
 import Sidebar, { SidebarItem, SidebarLink } from "./Sidebar/Sidebar"
-import Sidenav, {
-  SidenavHeader,
-  SidenavItem
-} from "./Sidenav/Sidenav"
+import Sidenav, { SidenavHeader, SidenavItem } from "./Sidenav/Sidenav"
 import Spinner from "./Spinner/Spinner"
 import Switch from "./Switch/Switch"
 import { Tabs, Tab } from "./Tabs/Tabs"
